chore: split tests to avoid stencil issue
